consolidate all individual Kessel services into one KesselInventoryService
