+++ api/kessel/inventory/v1beta2/inventory_service.proto
@@ -1,0 +1,57 @@
+syntax = "proto3";
+
+package kessel.inventory.v1beta2;
+
+import "google/api/annotations.proto";
+import "kessel/inventory/v1beta2/check_request.proto";
+import "kessel/inventory/v1beta2/check_response.proto";
+import "kessel/inventory/v1beta2/check_for_update_request.proto";
+import "kessel/inventory/v1beta2/check_for_update_response.proto";
+import "kessel/inventory/v1beta2/report_resource_request.proto";
+import "kessel/inventory/v1beta2/report_resource_response.proto";
+import "kessel/inventory/v1beta2/delete_resource_request.proto";
+import "kessel/inventory/v1beta2/delete_resource_response.proto";
+import "kessel/inventory/v1beta2/streamed_list_objects_request.proto";
+import "kessel/inventory/v1beta2/streamed_list_objects_response.proto";
+
+option go_package = "github.com/project-kessel/inventory-api/api/kessel/inventory/v1beta2";
+option java_multiple_files = true;
+option java_package = "org.project_kessel.api.inventory.v1beta2";
+
+service KesselInventoryService {
+  // Checks for the existence of a single Relationship
+  // (a Relation between a Resource and a Subject or Subject Set).
+  rpc Check(CheckRequest) returns (CheckResponse) {
+    option (google.api.http) = {
+      post: "/api/inventory/v1beta2/check"
+      body: "*"
+    };
+  }
+
+  rpc CheckForUpdate(CheckForUpdateRequest) returns (CheckForUpdateResponse) {
+    option (google.api.http) = {
+      post: "/api/inventory/v1beta2/checkforupdate"
+      body: "*"
+    };
+  }
+
+  rpc ReportResource(ReportResourceRequest) returns (ReportResourceResponse) {
+    option (google.api.http) = {
+      post: "/api/inventory/v1beta2/resources"
+      body: "*"
+    };
+  }
+
+  rpc DeleteResource(DeleteResourceRequest) returns (DeleteResourceResponse) {
+    option (google.api.http) = {
+      delete: "/api/inventory/v1beta2/resources"
+      body: "*"
+    };
+  }
+
+  rpc StreamedListObjects(StreamedListObjectsRequest) returns (stream StreamedListObjectsResponse) {
+    option (google.api.http) = {
+      get: "/api/inventory/v1beta2/resources"
+    };
+  }
+}
