+++ api/kessel/inventory/v1beta2/streamed_list_service.proto
@@ -2,18 +2,18 @@ syntax = "proto3";
 
 package kessel.inventory.v1beta2;
 
-import "google/api/annotations.proto";
+/*import "google/api/annotations.proto";
 import "kessel/inventory/v1beta2/streamed_list_objects_request.proto";
-import "kessel/inventory/v1beta2/streamed_list_objects_response.proto";
+import "kessel/inventory/v1beta2/streamed_list_objects_response.proto";*/
 
 option go_package = "github.com/project-kessel/inventory-api/api/kessel/inventory/v1beta2";
 option java_multiple_files = true;
 option java_package = "org.project_kessel.api.inventory.v1beta2";
 
 service KesselStreamedListService {
-  rpc StreamedListObjects(StreamedListObjectsRequest) returns (stream StreamedListObjectsResponse) {
+  /*rpc StreamedListObjects(StreamedListObjectsRequest) returns (stream StreamedListObjectsResponse) {
     option (google.api.http) = {
       get: "/api/inventory/v1beta2/resources"
     };
-  }
+  }*/
 }
