+++ internal/service/resources/kesselcheckservice.go
@@ -34,12 +34,6 @@ func NewKesselCheckServiceV1beta1(c *resources.Usecase) *KesselCheckServiceServi
 	}
 }
 
-func NewKesselCheckServiceV1beta2(c *resources.Usecase) *KesselCheckServiceServiceV1beta2 {
-	return &KesselCheckServiceServiceV1beta2{
-		Ctl: c,
-	}
-}
-
 func (s *KesselCheckServiceService) Check(ctx context.Context, req *pb.CheckRequest) (*pb.CheckResponse, error) {
 	identity, err := middleware.GetIdentity(ctx)
 	if err != nil {
@@ -58,32 +52,6 @@ func (s *KesselCheckServiceService) Check(ctx context.Context, req *pb.CheckRequ
 			},
 		}, *resource); err == nil {
 			return viewResponseFromAuthzRequest(resp), nil
-		} else {
-			return nil, err
-		}
-	} else {
-		return nil, err
-	}
-}
-
-func (s *KesselCheckServiceServiceV1beta2) Check(ctx context.Context, req *pbv1beta2.CheckRequest) (*pbv1beta2.CheckResponse, error) {
-	identity, err := middleware.GetIdentity(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	if resource, err := authzFromRequestV1beta2(identity, req.Object); err == nil {
-		if resp, err := s.Ctl.Check(ctx, req.GetRelation(), req.Object.Reporter.GetType(), &v1beta1.SubjectReference{
-			Relation: req.GetSubject().Relation,
-			Subject: &v1beta1.ObjectReference{
-				Type: &v1beta1.ObjectType{
-					Namespace: req.GetSubject().Resource.GetReporter().GetType(),
-					Name:      req.GetSubject().Resource.GetResourceType(),
-				},
-				Id: req.GetSubject().Resource.GetResourceId(),
-			},
-		}, *resource); err == nil {
-			return viewResponseFromAuthzRequestV1beta2(resp), nil
 		} else {
 			return nil, err
 		}
@@ -118,45 +86,10 @@ func (s *KesselCheckServiceService) CheckForUpdate(ctx context.Context, req *pb.
 	}
 }
 
-func (s *KesselCheckServiceServiceV1beta2) CheckForUpdate(ctx context.Context, req *pbv1beta2.CheckForUpdateRequest) (*pbv1beta2.CheckForUpdateResponse, error) {
-	identity, err := middleware.GetIdentity(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	if resource, err := authzFromRequestV1beta2(identity, req.Object); err == nil {
-		if resp, err := s.Ctl.CheckForUpdate(ctx, req.GetRelation(), req.Object.Reporter.GetType(), &v1beta1.SubjectReference{
-			Relation: req.GetSubject().Relation,
-			Subject: &v1beta1.ObjectReference{
-				Type: &v1beta1.ObjectType{
-					Namespace: req.GetSubject().Resource.GetReporter().GetType(),
-					Name:      req.GetSubject().Resource.GetResourceType(),
-				},
-				Id: req.GetSubject().Resource.GetResourceId(),
-			},
-		}, *resource); err == nil {
-			return updateResponseFromAuthzRequestV1beta2(resp), nil
-		} else {
-			return nil, err
-		}
-	} else {
-		return nil, err
-	}
-}
-
 func authzFromRequest(identity *authnapi.Identity, resource *pb.ObjectReference) (*model.ReporterResourceId, error) {
 	return &model.ReporterResourceId{
 		LocalResourceId: resource.Id,
 		ResourceType:    resource.Type.Name,
-		ReporterId:      identity.Principal,
-		ReporterType:    identity.Type,
-	}, nil
-}
-
-func authzFromRequestV1beta2(identity *authnapi.Identity, resource *pbv1beta2.ResourceReference) (*model.ReporterResourceId, error) {
-	return &model.ReporterResourceId{
-		LocalResourceId: resource.ResourceId,
-		ResourceType:    resource.ResourceType,
 		ReporterId:      identity.Principal,
 		ReporterType:    identity.Type,
 	}, nil
@@ -170,26 +103,10 @@ func viewResponseFromAuthzRequest(allowed bool) *pb.CheckResponse {
 	}
 }
 
-func viewResponseFromAuthzRequestV1beta2(allowed bool) *pbv1beta2.CheckResponse {
-	if allowed {
-		return &pbv1beta2.CheckResponse{Allowed: pbv1beta2.Allowed_ALLOWED_TRUE}
-	} else {
-		return &pbv1beta2.CheckResponse{Allowed: pbv1beta2.Allowed_ALLOWED_FALSE}
-	}
-}
-
 func updateResponseFromAuthzRequest(allowed bool) *pb.CheckForUpdateResponse {
 	if allowed {
 		return &pb.CheckForUpdateResponse{Allowed: pb.CheckForUpdateResponse_ALLOWED_TRUE}
 	} else {
 		return &pb.CheckForUpdateResponse{Allowed: pb.CheckForUpdateResponse_ALLOWED_FALSE}
-	}
-}
-
-func updateResponseFromAuthzRequestV1beta2(allowed bool) *pbv1beta2.CheckForUpdateResponse {
-	if allowed {
-		return &pbv1beta2.CheckForUpdateResponse{Allowed: pbv1beta2.Allowed_ALLOWED_TRUE}
-	} else {
-		return &pbv1beta2.CheckForUpdateResponse{Allowed: pbv1beta2.Allowed_ALLOWED_FALSE}
 	}
 }
